chores(files): Moves some components to their specific files
