+++ src/components/layout/auth-and-theme.tsx
@@ -1,0 +1,11 @@
+import AuthButton from './auth-button'
+import { ThemeToggle } from './theme-toggle'
+
+export default function AuthAndTheme() {
+  return (
+    <div className='flex items-center gap-x-2'>
+      <ThemeToggle />
+      <AuthButton />
+    </div>
+  )
+}
