+++ src/components/layout/links.ts
@@ -1,0 +1,21 @@
+export const LINKS: {
+  children: string
+  href: string
+  target?: string
+  rel?: string
+}[] = [
+  {
+    children: 'Posts',
+    href: '/posts'
+  },
+  {
+    children: 'Profile',
+    href: '/profile'
+  },
+  {
+    children: 'Github',
+    href: 'https://github.com/danybeltran/nextjs-typescript-and-mongodb',
+    target: '_blank',
+    rel: 'noreferrer'
+  }
+]
