+++ src/components/layout/navbar.tsx
@@ -1,139 +1,6 @@
 'use client'
-import Link from 'next/link'
 
-import { ThemeToggle, AuthButton } from '@/components/layout'
-import {
-  Button,
-  Dialog,
-  DialogClose,
-  DialogContent,
-  DialogDescription,
-  DialogTitle,
-  DialogTrigger
-} from '@/components/ui'
-
-import { usePathname } from 'next/navigation'
-import { cn, RenderList } from 'atomic-utils'
-import Icon from 'bs-icon'
-import { useState } from 'react'
-
-const NAVBAR_LINKS: {
-  children: string
-  href: string
-  target?: string
-  rel?: string
-}[] = [
-  {
-    children: 'Posts',
-    href: '/posts'
-  },
-  {
-    children: 'Profile',
-    href: '/profile'
-  },
-  {
-    children: 'Github',
-    href: 'https://github.com/danybeltran/nextjs-typescript-and-mongodb',
-    target: '_blank',
-    rel: 'noreferrer'
-  }
-]
-
-const useGetLinksStyle = () => {
-  const pathname = usePathname()
-  return function getLinkStyles(href: string) {
-    return cn('text-sm text-foreground/70 hover:text-foreground', {
-      'text-blue-700 dark:text-blue-400 font-medium hover:text-blue-700':
-        pathname.startsWith(href)
-    })
-  }
-}
-
-function AuthAndTheme() {
-  return (
-    <div className='flex items-center gap-x-2'>
-      <ThemeToggle />
-      <AuthButton />
-    </div>
-  )
-}
-
-function MobileMenu() {
-  const getLinkStyle = useGetLinksStyle()
-  const [isOpen, setIsOpen] = useState(false)
-
-  const hideMenu = () => setIsOpen(false)
-
-  return (
-    <div className='md:hidden'>
-      <Dialog open={isOpen} onOpenChange={setIsOpen}>
-        <DialogTrigger asChild>
-          <Button variant='ghost' size='icon' className='rounded-full'>
-            <Icon name='list' className='text-lg' />
-          </Button>
-        </DialogTrigger>
-        <DialogContent
-          className='h-full max-w-screen w-screen dark:bg-background/95 dark:backdrop-blur dark:supports-[backdrop-filter]:bg-background/60 rounded-none'
-          hideCloseButton
-        >
-          <DialogClose className='absolute left-6 top-3' asChild>
-            <Button size='icon' variant='ghost' className='rounded-full'>
-              <Icon name='x' className='text-lg' />
-            </Button>
-          </DialogClose>
-          <DialogTitle className='h-0 hidden'></DialogTitle>
-          <DialogDescription className='h-0 hidden'></DialogDescription>
-          <div className='flex flex-col items-center pt-16 gap-y-6'>
-            <Link
-              className='font-bold h-auto text-xl'
-              href={'/'}
-              onClick={hideMenu}
-            >
-              NEXT.JS
-            </Link>
-
-            <RenderList
-              data={NAVBAR_LINKS}
-              render={link => (
-                <Link
-                  key={'mobile' + link.href}
-                  onClick={hideMenu}
-                  className={cn(getLinkStyle(link.href), 'text-lg')}
-                  {...link}
-                />
-              )}
-            />
-            <AuthAndTheme />
-          </div>
-        </DialogContent>
-      </Dialog>
-    </div>
-  )
-}
-
-function DesktopMenu() {
-  const getLinkStyle = useGetLinksStyle()
-  return (
-    <>
-      <div className='space-x-3 hidden md:inline-block'>
-        <Link className='font-bold w-16 h-auto' href={'/'}>
-          NEXT.JS
-        </Link>
-        <RenderList
-          data={NAVBAR_LINKS}
-          render={link => (
-            <Link
-              key={'desktop' + link.href}
-              className={getLinkStyle(link.href)}
-              {...link}
-            />
-          )}
-        />
-      </div>
-      <AuthAndTheme />
-    </>
-  )
-}
+import { DesktopMenu, MobileMenu } from './menus'
 
 export default function Navbar() {
   return (
